group tests by need to restart OM
