+++ hadoop-ozone/integration-test/dev-support/findbugsExcludeFile.xml
@@ -98,7 +98,7 @@
     <Bug pattern="NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE" />
   </Match>
   <Match>
-    <Class name="org.apache.hadoop.ozone.om.TestOzoneManagerHAWithData"/>
+    <Class name="org.apache.hadoop.ozone.om.TestOzoneManagerHAWithStoppedNodes"/>
     <Bug pattern="DLS_DEAD_LOCAL_STORE" />
   </Match>
   <Match>
