--- hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestOzoneManagerHAMetadataOnly.java
@@ -1,590 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with this
- * work for additional information regarding copyright ownership.  The ASF
- * licenses this file to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package org.apache.hadoop.ozone.om;
-
-import org.apache.commons.lang3.RandomStringUtils;
-import org.apache.hadoop.hdds.protocol.StorageType;
-import org.apache.hadoop.hdfs.LogVerificationAppender;
-import org.apache.hadoop.ozone.ClientVersion;
-import org.apache.hadoop.ozone.OzoneTestUtils;
-import org.apache.hadoop.ozone.client.ObjectStore;
-import org.apache.hadoop.ozone.client.OzoneBucket;
-import org.apache.hadoop.ozone.client.BucketArgs;
-import org.apache.hadoop.ozone.client.OzoneVolume;
-import org.apache.hadoop.ozone.client.OzoneClient;
-import org.apache.hadoop.ozone.client.VolumeArgs;
-import org.apache.hadoop.ozone.om.exceptions.OMException;
-import org.apache.hadoop.ozone.om.ha.HadoopRpcOMFailoverProxyProvider;
-import org.apache.hadoop.ozone.om.ha.OMHAMetrics;
-import org.apache.hadoop.ozone.om.ha.OMProxyInfo;
-import org.apache.hadoop.ozone.om.helpers.OMRatisHelper;
-import org.apache.hadoop.ozone.om.ratis.OzoneManagerRatisServer;
-import org.apache.hadoop.ozone.om.request.volume.OMVolumeCreateRequest;
-import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos;
-import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos.CreateVolumeRequest;
-import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos.OMRequest;
-import org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos.VolumeInfo;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.ozone.test.GenericTestUtils;
-import org.apache.log4j.Logger;
-import org.apache.ratis.protocol.ClientId;
-import org.apache.ratis.protocol.Message;
-import org.apache.ratis.protocol.RaftClientReply;
-import org.apache.ratis.protocol.RaftClientRequest;
-import org.apache.ratis.server.RaftServer;
-import org.junit.Assert;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import javax.management.MBeanInfo;
-import javax.management.MBeanServer;
-import javax.management.ObjectName;
-import java.lang.management.ManagementFactory;
-import java.net.ConnectException;
-import java.net.InetSocketAddress;
-import java.time.Instant;
-import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.Iterator;
-import java.util.UUID;
-
-import static org.apache.hadoop.ozone.MiniOzoneHAClusterImpl.NODE_FAILURE_TIMEOUT;
-import static org.apache.hadoop.ozone.OzoneConfigKeys.OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT;
-
-import static org.apache.ratis.metrics.RatisMetrics.RATIS_APPLICATION_NAME_METRICS;
-import static org.junit.Assert.fail;
-
-/**
- * Test Ozone Manager Metadata operation in distributed handler scenario.
- */
-public class TestOzoneManagerHAMetadataOnly extends TestOzoneManagerHA {
-
-  private OzoneVolume createAndCheckVolume(String volumeName)
-      throws Exception {
-    String userName = "user" + RandomStringUtils.randomNumeric(5);
-    String adminName = "admin" + RandomStringUtils.randomNumeric(5);
-    VolumeArgs createVolumeArgs = VolumeArgs.newBuilder()
-        .setOwner(userName)
-        .setAdmin(adminName)
-        .build();
-
-    ObjectStore objectStore = getObjectStore();
-    objectStore.createVolume(volumeName, createVolumeArgs);
-
-    OzoneVolume retVolume = objectStore.getVolume(volumeName);
-
-    Assert.assertTrue(retVolume.getName().equals(volumeName));
-    Assert.assertTrue(retVolume.getOwner().equals(userName));
-    Assert.assertTrue(retVolume.getAdmin().equals(adminName));
-
-    return retVolume;
-  }
-
-  @Test
-  public void testAllVolumeOperations() throws Exception {
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-
-    createAndCheckVolume(volumeName);
-
-    ObjectStore objectStore = getObjectStore();
-    objectStore.deleteVolume(volumeName);
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.VOLUME_NOT_FOUND,
-        () -> objectStore.getVolume(volumeName));
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.VOLUME_NOT_FOUND,
-        () -> objectStore.deleteVolume(volumeName));
-  }
-
-
-  @Test
-  public void testAllBucketOperations() throws Exception {
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-    String bucketName = "volume" + RandomStringUtils.randomNumeric(5);
-
-    OzoneVolume retVolume = createAndCheckVolume(volumeName);
-
-    BucketArgs bucketArgs =
-        BucketArgs.newBuilder().setStorageType(StorageType.DISK)
-            .setVersioning(true).build();
-
-
-    retVolume.createBucket(bucketName, bucketArgs);
-
-
-    OzoneBucket ozoneBucket = retVolume.getBucket(bucketName);
-
-    Assert.assertEquals(volumeName, ozoneBucket.getVolumeName());
-    Assert.assertEquals(bucketName, ozoneBucket.getName());
-    Assert.assertTrue(ozoneBucket.getVersioning());
-    Assert.assertEquals(StorageType.DISK, ozoneBucket.getStorageType());
-    Assert.assertFalse(ozoneBucket.getCreationTime().isAfter(Instant.now()));
-
-
-    // Change versioning to false
-    ozoneBucket.setVersioning(false);
-
-    ozoneBucket = retVolume.getBucket(bucketName);
-    Assert.assertFalse(ozoneBucket.getVersioning());
-
-    retVolume.deleteBucket(bucketName);
-
-    OzoneTestUtils.expectOmException(OMException.ResultCodes.BUCKET_NOT_FOUND,
-        () -> retVolume.deleteBucket(bucketName));
-  }
-
-  /**
-   * Test that HadoopRpcOMFailoverProxyProvider creates an OM proxy
-   * for each OM in the cluster.
-   */
-  @Test
-  public void testOMProxyProviderInitialization() throws Exception {
-    OzoneClient rpcClient = getClient();
-
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-        OmFailoverProxyUtil.getFailoverProxyProvider(
-            rpcClient.getObjectStore().getClientProxy());
-
-    List<OMProxyInfo> omProxies =
-        omFailoverProxyProvider.getOMProxyInfos();
-
-    Assert.assertEquals(getNumOfOMs(), omProxies.size());
-
-    for (int i = 0; i < getNumOfOMs(); i++) {
-      InetSocketAddress omRpcServerAddr =
-          getCluster().getOzoneManager(i).getOmRpcServerAddr();
-      boolean omClientProxyExists = false;
-      for (OMProxyInfo omProxyInfo : omProxies) {
-        if (omProxyInfo.getAddress().equals(omRpcServerAddr)) {
-          omClientProxyExists = true;
-          break;
-        }
-      }
-      Assert.assertTrue("There is no OM Client Proxy corresponding to OM " +
-              "node" + getCluster().getOzoneManager(i).getOMNodeId(),
-          omClientProxyExists);
-    }
-  }
-
-  /**
-   * Test HadoopRpcOMFailoverProxyProvider failover on connection exception
-   * to OM client.
-   */
-  @Test
-  public void testOMProxyProviderFailoverOnConnectionFailure()
-      throws Exception {
-    ObjectStore objectStore = getObjectStore();
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-        OmFailoverProxyUtil
-            .getFailoverProxyProvider(objectStore.getClientProxy());
-    String firstProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    createVolumeTest(true);
-
-    // On stopping the current OM Proxy, the next connection attempt should
-    // failover to a another OM proxy.
-    getCluster().stopOzoneManager(firstProxyNodeId);
-    Thread.sleep(OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT * 4);
-
-    // Next request to the proxy provider should result in a failover
-    createVolumeTest(true);
-    Thread.sleep(OZONE_CLIENT_WAIT_BETWEEN_RETRIES_MILLIS_DEFAULT);
-
-    // Get the new OM Proxy NodeId
-    String newProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // Verify that a failover occurred. the new proxy nodeId should be
-    // different from the old proxy nodeId.
-    Assert.assertNotEquals("Failover did not occur as expected",
-        firstProxyNodeId, newProxyNodeId);
-  }
-
-  /**
-   * Test HadoopRpcOMFailoverProxyProvider failover when current OM proxy is not
-   * the current OM Leader.
-   */
-  @Test
-  public void testOMProxyProviderFailoverToCurrentLeader() throws Exception {
-    ObjectStore objectStore = getObjectStore();
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-            OmFailoverProxyUtil
-                    .getFailoverProxyProvider(objectStore.getClientProxy());
-
-    // Run couple of createVolume tests to discover the current Leader OM
-    createVolumeTest(true);
-    createVolumeTest(true);
-
-    // The oMFailoverProxyProvider will point to the current leader OM node.
-    String leaderOMNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // Perform a manual failover of the proxy provider to move the
-    // currentProxyIndex to a node other than the leader OM.
-    omFailoverProxyProvider.selectNextOmProxy();
-    omFailoverProxyProvider.performFailover(null);
-
-    String newProxyNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-    Assert.assertNotEquals(leaderOMNodeId, newProxyNodeId);
-
-    // Once another request is sent to this new proxy node, the leader
-    // information must be returned via the response and a failover must
-    // happen to the leader proxy node.
-    createVolumeTest(true);
-    Thread.sleep(2000);
-
-    String newLeaderOMNodeId =
-        omFailoverProxyProvider.getCurrentProxyOMNodeId();
-
-    // The old and new Leader OM NodeId must match since there was no new
-    // election in the Ratis ring.
-    Assert.assertEquals(leaderOMNodeId, newLeaderOMNodeId);
-  }
-
-  @Test
-  public void testOMRetryProxy() throws Exception {
-    int maxFailoverAttempts = getOzoneClientFailoverMaxAttempts();
-    // Stop all the OMs.
-    for (int i = 0; i < getNumOfOMs(); i++) {
-      getCluster().stopOzoneManager(i);
-    }
-
-    final LogVerificationAppender appender = new LogVerificationAppender();
-    final Logger logger = Logger.getRootLogger();
-    logger.addAppender(appender);
-
-    try {
-      createVolumeTest(true);
-      // After making N (set maxRetries value) connection attempts to OMs,
-      // the RpcClient should give up.
-      fail("TestOMRetryProxy should fail when there are no OMs running");
-    } catch (ConnectException e) {
-      Assert.assertEquals(1,
-          appender.countLinesWithMessage("Failed to connect to OMs:"));
-      Assert.assertEquals(maxFailoverAttempts,
-          appender.countLinesWithMessage("Trying to failover"));
-      Assert.assertEquals(1,
-          appender.countLinesWithMessage("Attempted " +
-              maxFailoverAttempts + " failovers."));
-    }
-  }
-
-  /**
-   * Choose a follower to send the request, the returned exception should
-   * include the suggested leader node.
-   */
-  @Test
-  public void testFailoverWithSuggestedLeader() throws Exception {
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-        OmFailoverProxyUtil
-            .getFailoverProxyProvider(getObjectStore().getClientProxy());
-
-    // Make sure All OMs are ready.
-    createVolumeTest(true);
-
-    // The OMFailoverProxyProvider will point to the current leader OM node.
-    String leaderOMNodeId = omFailoverProxyProvider.getCurrentProxyOMNodeId();
-    String leaderOMAddress = ((OMProxyInfo)
-        omFailoverProxyProvider.getOMProxyInfoMap().get(leaderOMNodeId))
-        .getAddress().getAddress().toString();
-    OzoneManager followerOM = null;
-    for (OzoneManager om: getCluster().getOzoneManagersList()) {
-      if (!om.isLeaderReady()) {
-        followerOM = om;
-        break;
-      }
-    }
-    assert followerOM != null;
-    Assertions.assertSame(followerOM.getOmRatisServer().checkLeaderStatus(),
-        OzoneManagerRatisServer.RaftServerStatus.NOT_LEADER);
-
-    OzoneManagerProtocolProtos.OMRequest writeRequest =
-        OzoneManagerProtocolProtos.OMRequest.newBuilder()
-            .setCmdType(OzoneManagerProtocolProtos.Type.ListVolume)
-            .setVersion(ClientVersion.CURRENT_VERSION)
-            .setClientId(UUID.randomUUID().toString())
-            .build();
-
-    try {
-      OzoneManagerProtocolProtos.OMResponse
-          omResponse = followerOM.getOmServerProtocol()
-          .submitRequest(null, writeRequest);
-      Assertions.fail("Test failure with NotLeaderException");
-    } catch (Exception ex) {
-      GenericTestUtils.assertExceptionContains("Suggested leader is OM:" +
-          leaderOMNodeId + "[" + leaderOMAddress + "]", ex);
-    }
-  }
-
-  @Test
-  public void testReadRequest() throws Exception {
-    String volumeName = "volume" + RandomStringUtils.randomNumeric(5);
-    ObjectStore objectStore = getObjectStore();
-    objectStore.createVolume(volumeName);
-
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-            OmFailoverProxyUtil
-                    .getFailoverProxyProvider(objectStore.getClientProxy());
-
-    String currentLeaderNodeId = omFailoverProxyProvider
-        .getCurrentProxyOMNodeId();
-
-    // A read request from any proxy should failover to the current leader OM
-    for (int i = 0; i < getNumOfOMs(); i++) {
-      // Failover omFailoverProxyProvider to OM at index i
-      OzoneManager ozoneManager = getCluster().getOzoneManager(i);
-
-      // Get the ObjectStore and FailoverProxyProvider for OM at index i
-      final ObjectStore store = getClient().getObjectStore();
-      final HadoopRpcOMFailoverProxyProvider proxyProvider =
-          OmFailoverProxyUtil.getFailoverProxyProvider(store.getClientProxy());
-
-      // Failover to the OM node that the objectStore points to
-      omFailoverProxyProvider.setNextOmProxy(
-          ozoneManager.getOMNodeId());
-
-      // A read request should result in the proxyProvider failing over to
-      // leader node.
-      OzoneVolume volume = store.getVolume(volumeName);
-      Assert.assertEquals(volumeName, volume.getName());
-
-      Assert.assertEquals(currentLeaderNodeId,
-          proxyProvider.getCurrentProxyOMNodeId());
-    }
-  }
-
-  @Test
-  public void testListVolumes() throws Exception {
-    String userName = UserGroupInformation.getCurrentUser().getUserName();
-    ObjectStore objectStore = getObjectStore();
-
-    String prefix = "vol-" + RandomStringUtils.randomNumeric(10) + "-";
-    VolumeArgs createVolumeArgs = VolumeArgs.newBuilder()
-        .setOwner(userName)
-        .setAdmin(userName)
-        .build();
-
-    Set<String> expectedVolumes = new TreeSet<>();
-    for (int i = 0; i < 100; i++) {
-      String volumeName = prefix + i;
-      expectedVolumes.add(volumeName);
-      objectStore.createVolume(volumeName, createVolumeArgs);
-    }
-
-    validateVolumesList(expectedVolumes,
-        objectStore.listVolumesByUser(userName, prefix, ""));
-
-    // Stop leader OM, and then validate list volumes for user.
-    stopLeaderOM();
-    Thread.sleep(NODE_FAILURE_TIMEOUT * 2);
-
-    validateVolumesList(expectedVolumes,
-        objectStore.listVolumesByUser(userName, prefix, ""));
-  }
-
-  @Test
-  public void testJMXMetrics() throws Exception {
-    // Verify any one ratis metric is exposed by JMX MBeanServer
-    OzoneManagerRatisServer ratisServer =
-        getCluster().getOzoneManager(0).getOmRatisServer();
-    ObjectName oname = new ObjectName(RATIS_APPLICATION_NAME_METRICS, "name",
-        RATIS_APPLICATION_NAME_METRICS + ".log_worker." +
-            ratisServer.getRaftPeerId().toString() +
-            "@" + ratisServer.getRaftGroup().getGroupId() + ".flushCount");
-    MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
-    MBeanInfo mBeanInfo = mBeanServer.getMBeanInfo(oname);
-    Assert.assertNotNull(mBeanInfo);
-    Object flushCount = mBeanServer.getAttribute(oname, "Count");
-    Assert.assertTrue((long) flushCount >= 0);
-  }
-
-  @Test
-  public void testOMRetryCache() throws Exception {
-    ObjectStore objectStore = getObjectStore();
-    objectStore.createVolume(UUID.randomUUID().toString());
-
-
-    HadoopRpcOMFailoverProxyProvider omFailoverProxyProvider =
-            OmFailoverProxyUtil
-                    .getFailoverProxyProvider(objectStore.getClientProxy());
-
-    String currentLeaderNodeId = omFailoverProxyProvider
-        .getCurrentProxyOMNodeId();
-
-    OzoneManagerRatisServer ozoneManagerRatisServer =
-        getCluster().getOzoneManager(currentLeaderNodeId).getOmRatisServer();
-
-    RaftServer raftServer = ozoneManagerRatisServer.getServer();
-
-    ClientId clientId = ClientId.randomId();
-    long callId = 2000L;
-    String userName = UserGroupInformation.getCurrentUser().getUserName();
-    String volumeName = UUID.randomUUID().toString();
-
-
-    GenericTestUtils.LogCapturer logCapturer = GenericTestUtils.LogCapturer
-        .captureLogs(OMVolumeCreateRequest.getLogger());
-
-    OzoneManagerProtocolProtos.UserInfo userInfo =
-        OzoneManagerProtocolProtos.UserInfo.newBuilder()
-            .setUserName("user")
-            .setHostName("host")
-            .setRemoteAddress("0.0.0.0")
-            .build();
-
-    OMRequest omRequest =
-        OMRequest.newBuilder().setCreateVolumeRequest(
-            CreateVolumeRequest.newBuilder().setVolumeInfo(
-                VolumeInfo.newBuilder().setOwnerName(userName)
-                    .setAdminName(userName).setVolume(volumeName).build())
-                .build()).setClientId(UUID.randomUUID().toString())
-            .setCmdType(OzoneManagerProtocolProtos.Type.CreateVolume)
-            .setUserInfo(userInfo)
-            .build();
-
-    RaftClientReply raftClientReply =
-        raftServer.submitClientRequest(RaftClientRequest.newBuilder()
-            .setClientId(clientId)
-            .setServerId(raftServer.getId())
-            .setGroupId(ozoneManagerRatisServer.getRaftGroup().getGroupId())
-            .setCallId(callId)
-            .setMessage(
-                Message.valueOf(
-                    OMRatisHelper.convertRequestToByteString(omRequest)))
-            .setType(RaftClientRequest.writeRequestType())
-            .build());
-
-    Assert.assertTrue(raftClientReply.isSuccess());
-
-    Assert.assertTrue(logCapturer.getOutput().contains("created volume:"
-        + volumeName));
-
-    logCapturer.clearOutput();
-
-    raftClientReply =
-        raftServer.submitClientRequest(RaftClientRequest.newBuilder()
-            .setClientId(clientId)
-            .setServerId(raftServer.getId())
-            .setGroupId(ozoneManagerRatisServer.getRaftGroup().getGroupId())
-            .setCallId(callId)
-            .setMessage(
-                Message.valueOf(
-                    OMRatisHelper.convertRequestToByteString(omRequest)))
-            .setType(RaftClientRequest.writeRequestType())
-            .build());
-
-    Assert.assertTrue(raftClientReply.isSuccess());
-
-    // As second time with same client id and call id, this request should
-    // not be executed ratis server should return from cache.
-    // If 2nd time executed, it will fail with Volume creation failed. check
-    // for that.
-    Assert.assertFalse(logCapturer.getOutput().contains(
-        "Volume creation failed"));
-
-    //Sleep for little above retry cache duration to get cache clear.
-    Thread.sleep(getRetryCacheDuration().toMillis() + 5000);
-
-    raftClientReply =
-        raftServer.submitClientRequest(RaftClientRequest.newBuilder()
-            .setClientId(clientId)
-            .setServerId(raftServer.getId())
-            .setGroupId(ozoneManagerRatisServer.getRaftGroup().getGroupId())
-            .setCallId(callId)
-            .setMessage(
-                Message.valueOf(
-                    OMRatisHelper.convertRequestToByteString(omRequest)))
-            .setType(RaftClientRequest.writeRequestType())
-            .build());
-
-    Assert.assertTrue(raftClientReply.isSuccess());
-
-    // As second time with same client id and call id, this request should
-    // be executed by ratis server as we are sending this request after cache
-    // expiry duration.
-    Assert.assertTrue(logCapturer.getOutput().contains(
-        "Volume creation failed"));
-
-  }
-
-  private void validateVolumesList(Set<String> expectedVolumes,
-      Iterator<? extends OzoneVolume> volumeIterator) {
-    int expectedCount = 0;
-
-    while (volumeIterator.hasNext()) {
-      OzoneVolume next = volumeIterator.next();
-      Assert.assertTrue(expectedVolumes.contains(next.getName()));
-      expectedCount++;
-    }
-
-    Assert.assertEquals(expectedVolumes.size(),  expectedCount);
-  }
-
-  @Test
-  public void testOMHAMetrics() throws Exception {
-    // Get leader OM
-    OzoneManager leaderOM = getCluster().getOMLeader();
-    // Store current leader's node ID,
-    // to use it after restarting the OM
-    String leaderOMId = leaderOM.getOMNodeId();
-    // Get a list of all OMs
-    List<OzoneManager> omList = getCluster().getOzoneManagersList();
-    // Check metrics for all OMs
-    checkOMHAMetricsForAllOMs(omList, leaderOMId);
-
-    // Restart leader OM
-    getCluster().shutdownOzoneManager(leaderOM);
-    getCluster().restartOzoneManager(leaderOM, true);
-    waitForLeaderToBeReady();
-
-    // Get the new leader
-    OzoneManager newLeaderOM = getCluster().getOMLeader();
-    String newLeaderOMId = newLeaderOM.getOMNodeId();
-    // Get a list of all OMs again
-    omList = getCluster().getOzoneManagersList();
-    // New state for the old leader
-    int newState = leaderOMId.equals(newLeaderOMId) ? 1 : 0;
-
-    // Get old leader
-    OzoneManager oldLeader = getCluster().getOzoneManager(leaderOMId);
-    // Get old leader's metrics
-    OMHAMetrics omhaMetrics = oldLeader.getOmhaMetrics();
-
-    Assertions.assertEquals(newState,
-        omhaMetrics.getOmhaInfoOzoneManagerHALeaderState());
-
-    // Check that metrics for all OMs have been updated
-    checkOMHAMetricsForAllOMs(omList, newLeaderOMId);
-  }
-
-  private void checkOMHAMetricsForAllOMs(List<OzoneManager> omList,
-      String leaderOMId) {
-    for (OzoneManager om : omList) {
-      // Get OMHAMetrics for the current OM
-      OMHAMetrics omhaMetrics = om.getOmhaMetrics();
-      String nodeId = om.getOMNodeId();
-
-      // If current OM is leader, state should be 1
-      int expectedState = nodeId
-          .equals(leaderOMId) ? 1 : 0;
-      Assertions.assertEquals(expectedState,
-          omhaMetrics.getOmhaInfoOzoneManagerHALeaderState());
-
-      Assertions.assertEquals(nodeId, omhaMetrics.getOmhaInfoNodeId());
-    }
-  }
-
-}
